feat: handle file descriptor in device
This change makes the Device struct take ownership of its File.
Devices created with libevdev require that their file descriptors
are not freed while the device is active. Now, the Device struct
will take ownership of the File, so that the user of the library
does not have to worry about this quirk.