rename fd methods to file, add UninitDevice
Methods called something fd or fd something are named confusingly, as
they often take `File`s as arguments. This was originally done to
mantain similarity with libevdev, but violates the least surprise
principle due to the existence of RawFd.

The UninitDevice struct was added to provide compatibility with
libevdev, which allows uninitialized devices, but mantains idiomatic
rust by encoding the fact that they are uninitalized in the type system.