rename fd methods to file, add UninitDevice
Methods called something fd or fd something are named confusingly, as
they often operate on `File`s. This was originally done to
mantain similarity with libevdev, but violates the least surprise
principle due to the existence of rust's `RawFd`.

The UninitDevice struct was added to provide compatibility with
libevdev, which allows uninitialized devices, but is idiomatic
rust, encoding the initalization in the type system.