evdev-sys: fix bug in build script
The build-script fails when building evdev-rs using `cargo package`.
As compared to `cargo build`, `cargo package` downloads all
dependencies from cargo registy, even for ones which are specified
using absolute path in Cargo.toml.

Due to this evdev-sys gets downloaded from cargo registry which
contains the libevdev submodule but it doesn't contain the full git
structure. This makes the build script believe that module is not
downloaded, but when it tries to excute the git command it is
unable to find a git directory.

Hence for this special case we need to ignore the failure while
downloading libevdev submodule.