Add type safe enable, diasble and has. (#70)
This commit adds trait Enable, which replaces the dynamic typing
previously used for enable, disable and has methods on DeviceWrapper.

To do so, it adds libevdev_property_disable to evdev-sys behind the
libevdev-1-10 feature. If this feature is enabled and the library is
found to be older than 1.10.0, build.rs will panic with a diagnostic
message. Additionally, a feature was added to evdev-rs which implements
Enable for InputProp. This requires the feature added to evdev-sys, so
will fail to compile if evdev-sys is too old.